apps/EirThings/requirements.txt
hx711==1.1.2.3
numpy==2.2.3
rpi-lgpio==0.6
python-dotenv==0.19.0
fastapi==0.115.8
uvicorn==0.34.0